requirements.txt
mysql-connector-python
ShopifyAPI
requests